Permissions update
Finished permissions section. Starting cart buildout next.

+++ backend/src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
 const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
 	async createItem(parent, args, ctx, info) {
@@ -42,9 +43,16 @@ const Mutations = {
 	async deleteItem(parent, args, ctx, info) {
 		const where = { id: args.id };
 		// 1. find the item
-		const item = await ctx.db.query.item({ where }, `{ id title}`);
+		const item = await ctx.db.query.item({ where }, `{ id title user { id}}`);
 		// 2. check if they own that item, or have the permissions
-		// TODO
+		const ownsItem = item.user.id === ctx.request.userId;
+		const hasPermissions = ctx.request.user.permissions.some((permission) =>
+			['ADMIN', 'ITEMDELETE'].includes(permission)
+		);
+
+		if (!ownsItem && !hasPermissions) {
+			throw new Error("You don't have permission to do that!");
+		}
 		// 3. delete it!
 		return ctx.db.mutation.deleteItem({ where }, info);
 	},
@@ -169,6 +177,37 @@ const Mutations = {
 		});
 		// 8. return the new user
 		return updatedUser;
+	},
+	async updatePermissions(parent, args, ctx, info) {
+		// 1. check if t hey are logged in
+		if (!ctx.request.userId) {
+			throw new Error('You must be logged in!');
+		}
+		// 2. query the current user
+		const currentUser = await ctx.db.query.user(
+			{
+				where: {
+					id: ctx.request.userId
+				}
+			},
+			info
+		);
+		// 3. check if they have permissions to do this
+		hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+		// 4. update the permissions
+		return ctx.db.mutation.updateUser(
+			{
+				data: {
+					permissions: {
+						set: args.permissions
+					}
+				},
+				where: {
+					id: args.userId
+				}
+			},
+			info
+		);
 	}
 };
